+++ app/src/main/java/com/nyx/mypurchases/data/products/ProductDaoRoomImpl.kt
@@ -24,7 +24,7 @@ interface ProductDaoRoomImpl : ProductDao {
     override fun updateProducts(products: List<ProductRoomEntity>)
 
     @Query("DELETE FROM product_table WHERE id=:productId")
-    override fun deleteProduct(productId: Long)
+    override fun deleteProduct(productId: Int)
 
     @Query("DELETE FROM product_table")
     override fun deleteProducts()
